+++ release_notes.md
@@ -5,6 +5,7 @@
 - Training status tasks upload training date and session to subject json
 - Query training criterion function added to brainbox.training
 - Swanson vector - option to annotate top/bottom 5 regions based on values, or pass in a list of acronyms
+- EphysQC can run without connection to ONE
 
 ## Release Notes 2.21
 ### Release Notes 2.21.3 2023-03-22