+++ release_notes.md
@@ -1,5 +1,5 @@
 ## Release Notes 2.22
-### Release Notes 2.22.0 (in preparation)
+### features
 - trials extractors support omissions trials from neuromodulator tasks
 - SpikeSortingLoader and EphysSessionLoader utils functions to load related objects such as drift
 - Training status tasks upload training date and session to subject json